+++ lib/core/api/endpoints/v3/auth_endpoint.dart
@@ -16,7 +16,9 @@ class AuthEndpoint {
 
   static String logout(String deviceToken) => "${_env.apiHost}/v1/auth/signout?device_token=$deviceToken";
 
-  static String recoverPassword = "${_env.apiHost}/v1/auth/forgot";
+  static String forgotPassword(String email) => "${_env.apiHost}/v3/auth/forgot-password/$email?isMobile=true";
+
+  static String resetPassword = "${_env.apiHost}/v3/auth/reset-password?isMobile=true";
 
   static String activity = "${_env.apiHost}/v1/auth/me/activity";