+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+# 0.6.2
+### Added
+- Adding forgot and reset password routes integration
+
+# 0.6.1
+### Added 
+- New functionalities to playlist repository
+
 # 0.6.0
 ### Added
 - Improving exceptions class by accepting code from backend